feat/ location of agencies in sub tasks with coordinates

+++ src/main/java/com/rescue/vscube/models/Coordinate.java
@@ -19,7 +19,7 @@ public class Coordinate {
     private double latitude;
 
     @JsonCreator
-    public Coordinate(@JsonProperty("latitude") float latitude, @JsonProperty("longitude") float longitude) {
+    public Coordinate(@JsonProperty("latitude") double latitude, @JsonProperty("longitude") double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
